requirements.txt
pygame==1.9.6
matplotlib==3.2.2
numpy==1.19.0
pillow==8.0.1
pyqlearning==1.2.5
pandas==1.1.4
keras==2.4.3
scipy==1.5.4